LPS-79679 Move to package .osgi.commands
